Recall and use region instance was launched in
When region configuration changes, the AWS provider loses track of the
launched instance because it attempts to find it by ID in the wrong
region. This commit addresses this issue by immediately saving the
region the instance was launched in into the machine's data directory
after the instance is launched. When the provider connects to AWS, it
will check if the machine's region state is already known, and will use
it if so, otherwise the connection will fallback to the default region
that is presently configured.