Get region value from provider.region, fix for a12ccea4
Commit a12ccea4 added code to persist the region associated with an
instance to use it later on when searching for the instance on AWS.
The modification added a function in the provider that returns the
region either from this saved state or from the default configuration.
This fix updates the code at various points to use this function rather
then rely on the default configuration for the region value.