Allow user to run post-termination tasks via callback
This commit adds a feature which allows the user to provide a Proc
callback to the VagrantPlugins::AWS::Config class. If set, then
immediately after the instance is terminated (in the destroy lifecycle),
the callback will get called. This facilitates use-cases where the user
needs to do post-destroy cleanup tasks, such as removing associated
Route53 records or anything else.